requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
python-decouple==3.8
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
